Server side SignalR Session work
